Start on printing citation table
NOTE: this function is not working as of this commit.

Idea is to take DNAbin object containing sequences with
'pubmed' attributes, and produce table with number of
papers from which sequences were taken, and summary information
for each.

In addition to not working, this is kind of hacky: treat it as a
first idea od how to do this!